isolate two types of genes
