+++ inc/genotype.hpp
@@ -8,40 +8,11 @@
 #include <cstdint>
 #include <utility>
 #include <filesystem>
+#include "gene.hpp"
 #include "graph-network.hpp"
 
 // using declarations
 using std::uint64_t;
-using std::int64_t;
-
-// define three node types (hidden, input(sensor), output)
-enum struct NodeType{
-        hidden,
-        sensor,
-        output
-};
-
-// define node genes
-struct Node{
-        uint64_t node_number;
-        NodeType node_type;
-
-        // return the char representation of node type (when printing the node types)
-        static char get_nodetype(const NodeType type) noexcept;
-        // from the char representation to node type enum
-        static NodeType get_nodetype(const char type) noexcept;
-};
-
-// define connection genes
-struct Connection{
-        uint64_t in, out;
-        long double weight;
-        bool enable;
-        uint64_t innov;
-
-        // return the string representation of each connection (when printing connections)
-        static std::string make_connect(const Connection& connect);
-};
 
 // define genotype
 class Genotype{
