+++ src/genotype.cpp
@@ -11,27 +11,6 @@
 #include <iostream>
 #include <algorithm>
 
-char Node::get_nodetype(const NodeType type) noexcept{
-        if(type == NodeType::hidden) return 'H';
-        else if(type == NodeType::output) return 'O';
-        return 'S';
-}
-
-NodeType Node::get_nodetype(const char type) noexcept{
-        if(type == 'H') return NodeType::hidden;
-        else if(type == 'O') return NodeType::output;
-        return NodeType::sensor;
-}
-
-std::string Connection::make_connect(const Connection& connect){
-        std::ostringstream connection;
-        connection << connect.in << ' ' << connect.out << ' ';
-        connection << std::setprecision(std::numeric_limits<long double>::max_digits10) << connect.weight << ' ';
-        connection << (connect.enable ? 'E' : 'D') << ' ';
-        connection << connect.innov;
-        return connection.str();
-}
-
 // this constructor creates a network with no hidden nodes
 // inputs and outputs forms a fully connected graph, each edge receives a weight of 1;
 Genotype::Genotype(const int inputs, const int outputs){
